fix(storage): resolve NVS key length limit and memory leaks
- Removed redundant MAC address from softIron/hardIron keys in Sub to keep under 15-char limit.
- Fixed type mismatch in Sub memBuoyId (changed from int8_t/getChar to uint64_t/getULong64).
- Added missing storage.end() to initMemory in Top and Sub to prevent unclosed namespaces.

+++ Firmware/RobobuoySub/src/datastorage.h
@@ -7,7 +7,7 @@
 
 void initMemory(void);
 void memDockPos(RoboStruct *buoy, bool get);
-void memBuoyId(int8_t *id, bool get);
+void memBuoyId(uint64_t *id, bool get);
 void CompassCallibrationFactors(RoboStruct *buoy, bool get);
 void Declination(RoboStruct *buoy, bool get);
 void CompasOffset(RoboStruct *buoy, bool get);
